+++ src/bsgo/components/ResourceComponent.cc
@@ -3,7 +3,7 @@
 
 namespace bsgo {
 
-ResourceComponent::ResourceComponent(const Uuid resource, const float amount)
+ResourceComponent::ResourceComponent(const Uuid resource, const int amount)
   : AbstractComponent(ComponentType::RESOURCE)
   , m_resource(resource)
   , m_amount(amount)
@@ -14,14 +14,14 @@ auto ResourceComponent::resource() const -> Uuid
   return m_resource;
 }
 
-auto ResourceComponent::amount() const -> float
+auto ResourceComponent::amount() const -> int
 {
   return m_amount;
 }
 
 void ResourceComponent::update(const float /*elapsedSeconds*/) {}
 
-void ResourceComponent::setAmount(const float amount)
+void ResourceComponent::setAmount(const int amount)
 {
   if (amount < 0)
   {
